feat: add parameter -f to service apply, service delete, service open and service verify

+++ src/serviceDefinition/delete.ts
@@ -6,19 +6,28 @@ import { loadServiceDefinition } from './files';
 import { abortExecution } from '../errors';
 import { executeApiCall } from '../api';
 
-export async function deleteServiceDefinition(serviceDefinitionPathOrId: string) {
-  let id = serviceDefinitionPathOrId;
-  try {
-    const serviceDefinition = await loadServiceDefinition(serviceDefinitionPathOrId);
-    id = serviceDefinition.id ?? id;
-  } catch (e) {
-    // swallow silently assuming that the parameter is actually the ID and not a file path
+export interface Options {
+  id?: string;
+  file: string;
+}
+
+export async function deleteServiceDefinition(options: Options) {
+  if(!options.id && !options.file){
+    throw abortExecution(
+      'Failed to delete service definition, no id nor path to the service definition file was given',
+    );
   }
+  let id = options.id;
+  if(!options.id && options.file){
+    const serviceDefinition = await loadServiceDefinition(options.file);
+    id = serviceDefinition.id ?? id;
+  }
+  // assuming that the parameter id is given
 
   try {
     await executeApiCall({
       method: 'DELETE',
-      path: `/api/service-definitions/${encodeURIComponent(id)}`,
+      path: `/api/service-definitions/${encodeURIComponent(String(id))}`,
     });
   } catch (e) {
     throw abortExecution(
